+++ R/queries.R
@@ -13,6 +13,13 @@ setGeneric(
   }
 )
 setGeneric(
+  "involvedVariables", function(x) {
+    stop(
+      "No available application of `involvedVariables` for this object."
+    )
+  }
+)
+setGeneric(
   "getCoefficient", function(qspray, exponents) {
     stop(
       "No available application of `getCoefficient`. Check the arguments."
@@ -59,12 +66,14 @@ setGeneric(
 #' @aliases numberOfVariables,qspray-method 
 #' @docType methods
 #' @title Number of variables in a 'qspray' polynomial
-#' @description Number of variables involved in a \code{qspray} object.
+#' @description Number of variables involved in a \code{qspray} object (see 
+#'   the note for the precise meaning).
 #'
 #' @param x a \code{qspray} object
 #'
 #' @return An integer.
 #' @export
+#' @seealso \code{\link{involvedVariables}}.
 #' @note The number of variables in the \code{qspray} object \code{y^2+1} where
 #'   \code{y=qlone(2)} is \code{2}, not \code{1}, although only one variable is 
 #'   present. Strictly speaking, the function returns the maximal integer 
@@ -76,11 +85,40 @@ setMethod(
   }
 )
 
+#' @name involvedVariables
+#' @aliases involvedVariables,qspray-method 
+#' @docType methods
+#' @title Variables involved in a 'qspray' polynomial
+#' @description Variables involved in a \code{qspray} object.
+#'
+#' @param x a \code{qspray} object
+#'
+#' @return A vector of integers. Each integer represents the index of a 
+#'   variable involved in \code{x}.
+#' @export
+#' @examples
+#' x <- qlone(1); z <- qlone(3)
+#' involvedVariables(x^2 + x*z + 1) # should be c(1L, 3L)
+setMethod(
+  "involvedVariables", "qspray", 
+  function(x) {
+    if(isConstant(x)) {
+      integer(0L)
+    } else {
+      M <- powersMatrix(x)
+      tests <- apply(M, 2L, function(col) {
+        any(col != 0L)
+      })
+      which(tests)
+    }
+  }
+)
+
 #' @name numberOfTerms
 #' @aliases numberOfTerms,qspray-method 
 #' @docType methods
 #' @title Number of terms in a 'qspray' polynomial
-#' @description Number of terms in the polynomial defined by a 
+#' @description Number of terms of the polynomial defined by a 
 #'   \code{qspray} object.
 #'
 #' @param qspray a \code{qspray} object
@@ -98,11 +136,11 @@ setMethod(
 #' @aliases getCoefficient,qspray,numeric-method 
 #' @docType methods
 #' @title Get a coefficient in a 'qspray' polynomial
-#' @description Get the coefficient corresponding to the given sequence of 
-#'   exponents.
+#' @description Get the coefficient of the term with the given monomial.
 #' 
 #' @param qspray a \code{qspray} object
-#' @param exponents a vector of exponents
+#' @param exponents a vector of exponents, thereby defining a monomial;
+#'   trailing zeros are ignored
 #'
 #' @return The coefficient as a \code{bigq} number.
 #' @export
@@ -113,12 +151,12 @@ setMethod(
 #' x <- qlone(1)
 #' y <- qlone(2)
 #' p <- 4*x^2 + 3*y - 5
-#' getCoefficient(p, 2)
-#' getCoefficient(p, c(2, 0)) # same as getCoefficient(p, 2)
-#' getCoefficient(p, c(0, 1)) # coefficient of y because y=x^0*y^1
-#' getCoefficient(p, 0) # the constant term
+#' getCoefficient(p, 2)          # coefficient of x^2
+#' getCoefficient(p, c(2, 0))    # same as getCoefficient(p, 2)
+#' getCoefficient(p, c(0, 1))    # coefficient of y because y=x^0*y^1
+#' getCoefficient(p, 0)          # the constant term
 #' getCoefficient(p, integer(0)) # the constant term 
-#' getCoefficient(p, 3) # there's no cubic term
+#' getCoefficient(p, 3)          # there's no x^3
 setMethod(
   "getCoefficient", c("qspray", "numeric"), 
   function(qspray, exponents) {
@@ -213,7 +251,7 @@ setMethod(
 setMethod(
   "isQzero", "qspray", 
   function(qspray) {
-    arity(qspray) == -Inf
+    length(qspray@coeffs) == 0L
   }
 )
 
@@ -231,7 +269,7 @@ setMethod(
 setMethod(
   "isQone", "qspray", 
   function(qspray) {
-    isConstant(qspray) && getConstantTerm(qspray) == 1L
+    length(qspray@coeffs) == 1L && getConstantTerm(qspray) == 1L
   }
 )
 
@@ -276,4 +314,38 @@ collinearQsprays <- function(qspray1, qspray2) {
   }
   r <- coeffs2[1L] / coeffs1[1L]
   all(r * coeffs1 == coeffs2)
+}
+
+#' @title Whether a 'qspray' polynomial is homogeneous
+#' @description Checks whether the polynomial defined by a \code{qspray} 
+#'   object is homogeneous, and also returns the degree if this is true.
+#'
+#' @param qspray a \code{qspray} object
+#'
+#' @return A Boolean value indicating whether the polynomial defined by 
+#'   \code{qspray} is homogeneous. Moreover, if it is homogeneous, the degree 
+#'   is given in the attribute \code{"degree"} of the output.
+#' @export
+#'
+#' @examples
+#' lambda <- c(3, 2, 1)
+#' p <- PSFpoly(4, lambda)
+#' ( homogeneous <- isHomogeneousQspray(p) ) # should be TRUE
+#' attr(homogeneous, "degree") == sum(lambda) # should be TRUE
+isHomogeneousQspray <- function(qspray) {
+  if(isConstant(qspray)) {
+    out <- TRUE
+    attr(out, "degree") <- 0L
+  } else if(getConstantTerm(qspray) != 0L) {
+    out <- FALSE
+  } else {
+    degrees <- vapply(qspray@powers, sum, integer(1L))
+    if(all(degrees == degrees[1L])) {
+      out <- TRUE
+      attr(out, "degree") <- degrees[1L]
+    } else {
+      out <- FALSE
+    }
+  }
+  out
 }