+++ R/transformation.R
@@ -132,6 +132,9 @@ setMethod(
   "permuteVariables", c("qspray", "numeric"), 
   function(x, permutation) {
     stopifnot(isPermutation(permutation))
+    if(isConstant(x)) {
+      return(x)
+    }
     m <- numberOfVariables(x)
     n <- length(permutation)
     if(m > n) {
@@ -142,7 +145,7 @@ setMethod(
     for(. in seq_len(n - m)) {
       M <- cbind(M, 0L)
     }
-    M <- M[, permutation]
+    M <- M[, permutation, drop = FALSE]
     powers <- apply(M, 1L, function(row) {
       removeTrailingZeros(row)
     }, simplify = FALSE)
@@ -186,6 +189,9 @@ setMethod(
   "swapVariables", c("qspray", "numeric", "numeric"), 
   function(x, i, j) {
     stopifnot(isNonnegativeInteger(i), isNonnegativeInteger(j))
+    if(isConstant(x)) {
+      return(x)
+    }
     m <- numberOfVariables(x)
     n <- max(m, i, j)
     permutation <- seq_len(n)
@@ -195,7 +201,7 @@ setMethod(
     for(. in seq_len(n - m)) {
       M <- cbind(M, 0L)
     }
-    M <- M[, permutation]
+    M <- M[, permutation, drop = FALSE]
     powers <- apply(M, 1L, function(row) {
       removeTrailingZeros(row)
     }, simplify = FALSE)
@@ -228,13 +234,16 @@ setMethod(
 #' Y <- x + y + 1
 #' composeQspray(P, list(X, Y)) # this is P(x^2, x+y+1)
 composeQspray <- function(qspray, listOfQsprays) {
+  if(isConstant(qspray)) {
+    return(qspray)
+  }
   n <- numberOfVariables(qspray)
   if(length(listOfQsprays) < n) {
     stop(
       sprintf(
         paste0(
           "The `listOfQsprays` argument must be a list containing ", 
-          "at least %d objects coercable to `qspray` polynomials."
+          "at least %d objects coercible to `qspray` polynomials."
         ), n
       )
     )
@@ -252,6 +261,10 @@ composeQspray <- function(qspray, listOfQsprays) {
       }
     }
     result <- result + coeffs[i] * term
+  }
+  if(isNamedList(listOfQsprays)) {
+    showQsprayOption(result, "showQspray") <- 
+      showQsprayXYZ(names(listOfQsprays))
   }
   result  
 }
@@ -275,8 +288,10 @@ setGeneric(
 #' 
 #' @param x a \code{qspray} polynomial
 #' @param listOfQsprays a list containing at least \code{n} \code{qspray} 
-#'   objects, or objects coercable to \code{qspray} objects, where \code{n} 
-#'   is the number of variables of the polynomial given in the \code{x} argument
+#'   objects, or objects coercible to \code{qspray} objects, where \code{n} 
+#'   is the number of variables of the polynomial given in the \code{x} 
+#'   argument; if this list is named, then its names will be used in the
+#'   show options of the result
 #'
 #' @return The \code{qspray} polynomial obtained by replacing the variables of 
 #'   the polynomial given in the \code{x} argument with the polynomials given